Get rid of EventEmitter pattern in favor of Redux
Redux has proven a better pattern to manage the data flow.
This is a first step towards the goal of creating an adapter
between SocketIO events and Redux actions to be reused by any host.
This would help us to improve testability and readability.
As a side-effect we also reduce bundle size.

We now provide the auth as a function that returns a Promise
indicating whether the authentication mechanism was successful or not.

Note that wpcom specific code (how to get userId, locale, groups)
is separated from the connection, so the data necessary for Happychat
is made more evident.